emitter/requirements.txt
boto3==1.34.69
pyyaml==6.0.1
